feat!: re-implement server with Connect RPC (#35)
This change overhauls the server and API to use [Connect
RPC](https://connectrpc.com/). This provides us with the ability to
easily generate client libraries with our Protobuf definition, and will
make developing the companion command line interface much easier.

While switching to Connect RPC is absolutely a breaking change, it also
provides a nice REST API that is aligned with our gRPC service
definitions.

Release-As: 0.10.0